Move the BottomSheetBackground to its own component
